skills/unifai-trading-suite/pyproject.toml
[project]
name = "ai-trader-prediction-market"
version = "0.1.0"
description = "AI-powered trading agent for prediction markets"
requires-python = ">=3.10"
dependencies = [
    "unifai-sdk>=0.3.3",
    "litellm>=1.40.0",
    "google-generativeai>=0.5.0",
    "aiohttp>=3.9.0",
    "httpx>=0.27.0",
    "web3>=6.15.0",
    "pydantic>=2.6.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.3.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
